# backend/docs/swagger.yaml
definitions:
...
  backend_api_v1.ForgotPasswordRequest:
    properties:
      email:
        description: 邮箱
        example: zhangsan@example.com
        type: string
    required:
    - email
    type: object
...
  backend_api_v1.LoginRequest:
    properties:
      autoLogin:
        description: 记住我 - 延长refresh token有效期
        type: boolean
      password:
        description: 密码
        example: "123456"
        type: string
      username:
        description: 用户名 或 邮箱
        example: zhangsan
        type: string
    required:
    - password
    - username
    type: object
  backend_api_v1.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/backend_api_v1.TokenData'
      errorCode:
        description: 错误码
        type: integer
      errorMessage:
        description: 报错信息
        type: string
      errorShowType:
        description: 前端展示方式
        type: integer
      success:
        description: 是否成功
        example: true
        type: boolean
    type: object
  backend_api_v1.LogoutData:
    properties:
      endSessionUrl:
        description: OIDC end session URL for SLO
        type: string
    type: object
  backend_api_v1.LogoutResponse:
    properties:
      data:
        $ref: '#/definitions/backend_api_v1.LogoutData'
      errorCode:
        description: 错误码
        type: integer
      errorMessage:
        description: 报错信息
        type: string
      errorShowType:
        description: 前端展示方式
        type: integer
      success:
        description: 是否成功
        example: true
        type: boolean
    type: object
...
  backend_api_v1.OIDCAuthRequest:
    properties:
      autoLogin:
        description: 记住我 - 延长refresh token有效期
        type: boolean
      code:
        type: string
      state:
        type: string
    required:
    - code
    type: object
...
  backend_api_v1.RefreshTokenRequest:
    properties:
      refreshToken:
        description: 刷新凭证
        type: string
    required:
    - refreshToken
    type: object
  backend_api_v1.RegisterRequest:
    properties:
      email:
        description: 邮箱
        example: zhangsan@example.com
        type: string
      fullName:
        description: 全名
        example: Zhang San
        type: string
      password:
        description: 密码
        example: "123456"
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  backend_api_v1.ResetPasswordRequest:
    properties:
      newPassword:
        description: 新密码
        example: newpass123
        minLength: 6
        type: string
      token:
        description: 重置令牌
        type: string
    required:
    - newPassword
    - token
    type: object
  backend_api_v1.Response:
    properties:
      data:
        description: 返回数据
      errorCode:
        description: 错误码
        type: integer
      errorMessage:
        description: 报错信息
        type: string
      errorShowType:
        description: 前端展示方式
        type: integer
      success:
        description: 是否成功
        example: true
        type: boolean
    type: object
...
  backend_api_v1.TokenData:
    properties:
      accessToken:
        description: 访问凭证
        type: string
      expiresIn:
        description: 有效期（秒）
        type: integer
      refreshToken:
        description: 刷新凭证
        type: string
      tokenType:
        description: 凭证类型
        type: string
    type: object
...
host: localhost:7001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vibe Testing
  version: 1.0.0
paths:
...
  /auth/oidc:
    post:
      consumes:
      - application/json
      description: 通过OIDC授权码登录
      operationId: OIDCAuth
      parameters:
      - description: OIDC授权码
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.OIDCAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.LoginResponse'
      summary: OIDC登录
      tags:
      - Auth
  /forgot-password:
    post:
      consumes:
      - application/json
      description: 发送重置密码邮件
      operationId: ForgotPassword
      parameters:
      - description: 忘记密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      summary: 忘记密码
      tags:
      - Auth
...
  /login:
    post:
      consumes:
      - application/json
      description: 支持用户名或邮箱登录，LDAP启用时自动尝试LDAP认证
      operationId: Login
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.LoginResponse'
      summary: 登录
      tags:
      - Auth
  /logout:
    post:
      consumes:
      - application/json
      description: 用户登出，记录审计日志
      operationId: Logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.LogoutResponse'
      security:
      - Bearer: []
      summary: 登出
      tags:
      - Auth
  /refresh-token:
    post:
      consumes:
      - application/json
      description: 刷新访问令牌和刷新令牌
      operationId: RefreshToken
      parameters:
      - description: 刷新令牌信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.LoginResponse'
      summary: 刷新令牌
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: 目前只支持通过邮箱进行注册
      operationId: Register
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      summary: 注册
      tags:
      - Auth
  /reset-password:
    post:
      consumes:
      - application/json
      description: 通过token重置密码
      operationId: ResetPassword
      parameters:
      - description: 重置密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      summary: 重置密码
      tags:
      - Auth
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"